Move play state parser so it's available in tournament tests
